+++ NEWS.md
@@ -1,3 +1,5 @@
+# simstudy (development version)
+
 # simstudy (0.9.2)
 
 * A minor fix in a unit test for correlated matrices that was failing on CRAN due to